requirements.txt
fastapi
uvicorn[standard]
transformers
Pillow
python-multipart